requirements.txt
flask
SQLAlchemy
psycopg2-binary
celery
redis
flask-mail
python-dotenv